+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `@shgysk8zer0/js-utils`
+
+### Removed
+- Remove `eslint` and `rollup`
+
+### Changed
+- Update rollup config
+
 ## [v1.0.2] - 2023-06-01
 
 ### Fixed